move models into subdir
