+++ crypto-scraper/app/Console/Commands/Bitcointalk/LoadMainTopics.php
@@ -4,7 +4,7 @@ namespace App\Console\Commands\Bitcointalk;
 
 use App\Console\BitcointalkParser;
 use App\Models\Pg\Bitcointalk\MainTopic;
-use App\Models\Pg\BoardPage;
+use App\Models\Pg\Bitcointalk\BoardPage;
 
 class LoadMainTopics extends BitcointalkParser {
     const ENTITY = 'topic';
